update curl library tests

+++ tests/include/lib/src/CurlManager.php
@@ -14,7 +14,7 @@ class CurlManager
         return "http://127.0.0.1:{$this->port}";
     }
 
-    protected function run_cli_server($port)
+    protected function runCliServer($port)
     {
         $proc = new Process(function (Process $p) use ($port) {
             $exec = "/usr/bin/env php -t " . __DIR__ . " -n -S 127.0.0.1:{$port} " . __DIR__ . "/responder/get.php";
@@ -22,7 +22,7 @@ class CurlManager
         }, true, 1);
 
         $proc->start();
-        while (true) {
+        while (1) {
             usleep(10000);
             if (@file_get_contents($this->getUrlBase() . '/')) {
                 break;
@@ -31,10 +31,14 @@ class CurlManager
         return $proc;
     }
 
-    function run(callable $fn)
+    function run(callable $fn, $createCliServer = true)
     {
-        $this->port = get_one_free_port();
-        $proc = $this->run_cli_server($this->port);
+        if ($createCliServer) {
+            $this->port = get_one_free_port();
+            $proc = $this->runCliServer($this->port);
+        } else {
+            $proc = null;
+        }
 
         global $argc, $argv;
         if ($argc > 1 and $argv[1] == 'ori') {
@@ -44,9 +48,14 @@ class CurlManager
         }
 
         go(function () use ($fn, $proc) {
-            $fn($this->getUrlBase());
-            Swoole\Process::kill($proc->pid);
+            $fn("127.0.0.1:{$this->port}");
+            if ($proc) {
+                Swoole\Process::kill($proc->pid);
+            }
         });
         Swoole\Event::wait();
+        if ($createCliServer) {
+            Process::wait();
+        }
     }
 }
